tg_ai_bot/requirements.txt
requests==2.31.0 
python-telegram-bot
ollama
